Made changes to the homepage of the blog section. since you guys were just bashing me here and there
Made changes to the homepage of the blog section. since you guys were just bashing me here and there. change it from the stretched images to a small image on the left and the text beside it at the right hand side. Also removed the yellow line and also changed the font color from green to black as you guys requested